+++ lib/deimos.rb
@@ -49,7 +49,7 @@ module Deimos
     deimos.encode_message
     deimos.batch_consumption.invalid_records
     deimos.batch_consumption.valid_records
-    deimos.batch_consumption.individual_fallback
+    deimos.batch_consumption.initial_failure
     deimos.outbox.produce
   ).freeze